refator : AuthUser 반영 #26

+++ airbnbClone/src/main/java/dbDive/airbnbClone/api/reservation/service/ReservationService.java
@@ -7,6 +7,7 @@ import dbDive.airbnbClone.api.reservation.dto.request.BookRequest;
 import dbDive.airbnbClone.api.reservation.dto.response.HostTotalAccommodationResponse;
 import dbDive.airbnbClone.api.reservation.dto.response.TotalAccommodationResponse;
 import dbDive.airbnbClone.common.GlobalException;
+import dbDive.airbnbClone.config.auth.AuthUser;
 import dbDive.airbnbClone.entity.accommodation.Accommodation;
 import dbDive.airbnbClone.entity.resevation.Reservation;
 import dbDive.airbnbClone.repository.accommodation.AccommodationRepository;
@@ -26,7 +27,7 @@ public class ReservationService {
     private final ReservationRepository reservationRepository;
 
     @Transactional
-    public void bookAccommodation(Long accommodationId, BookRequest request) {
+    public void bookAccommodation(Long accommodationId, BookRequest request, AuthUser authUser) {
 
         //숙소 예약 -> 프론트에서  checkIn/checkOut/totalPrice/guest 받아서 예약 진행
         Accommodation findAcmd = accommodationRepository.findById(accommodationId).orElseThrow(() -> new GlobalException("존재하지 않는 숙소입니다,"));
@@ -35,32 +36,40 @@ public class ReservationService {
                 .checkOut(request.getCheckOut())
                 .totalPrice(request.getTotalPrice())
                 .guest(request.getGuest())
+                .user(authUser.getUser())
                 .accommodation(findAcmd)
                 .build();
         reservationRepository.save(reservation);
     }
 
-    public TotalAccommodationResponse allAccommodations(Pageable pageable) {
+    public TotalAccommodationResponse allAccommodations(Pageable pageable, AuthUser authUser) {
 
-        PageImpl<ReservationDto> result = reservationRepository.findAllReservations(pageable);
+        PageImpl<ReservationDto> result = reservationRepository.findAllReservations(pageable, authUser);
         return new TotalAccommodationResponse(result);
 
     }
 
-    public SelectReservationDto selectAccommodations(Long reservationId) {
+    public SelectReservationDto selectAccommodations(Long reservationId, AuthUser authUser) {
         // 게스트 - 숙소 예약 단건 조회
-        SelectReservationDto selectReservation = reservationRepository.findSelectReservation(reservationId);
+        SelectReservationDto selectReservation = reservationRepository.findSelectReservation(reservationId, authUser);
         return selectReservation;
 
     }
 
-    public void deleteAccommodation(Long reservationId) {
+    public void deleteAccommodation(Long reservationId, AuthUser authUser) {
         // 게스트 - 예약 취소
         Reservation findReservation = reservationRepository.findById(reservationId).orElseThrow();
-        reservationRepository.delete(findReservation);
+        Long reservationUserId = findReservation.getUser().getId();
+        Long authUserId = authUser.getUser().getId();
+
+        if (reservationUserId.equals(authUserId)) {
+            reservationRepository.delete(findReservation);
+        } else {
+            throw new GlobalException("예약 취소 권한이 없습니다.");
+        }
     }
 
-    public HostTotalAccommodationResponse hostAllAccommodations(Long userId, Pageable pageable) {
+    public HostTotalAccommodationResponse hostAllAccommodations(Pageable pageable, AuthUser authUser) {
         // 호스트 - 예약 조회
         /**
          * 1. 등록된 accommodation에서 userId로 accommodationId 찾기
@@ -71,7 +80,7 @@ public class ReservationService {
          * - isDeleted == true, 취소
          */
 
-        PageImpl<HostReservationDto> hostReservation = reservationRepository.findHostReservation(userId, pageable);
+        PageImpl<HostReservationDto> hostReservation = reservationRepository.findHostReservation(pageable, authUser);
         return new HostTotalAccommodationResponse(hostReservation);
     }
